[FIX] The list and remove functions have been edited so that they block if they change idAdmin to true   modified: app/controllers/dailyNoteController.js  Mofified: app/controller/helpCenterAskController.js  Modifies: app/controller/helpCenterController.js

+++ app/controllers/helpCenterAskController.js
@@ -67,31 +67,60 @@ class HelperCenterAskController extends Controller {
                     console.log(error)
                     return resp.status(400).send(JSON.stringify({ erro: 'USUARIO não existente' }));
                 }
-                userDao.checkAdmin(email, (err, docs) => {
-                    // console.log(docs.isAdmin);
+                const helperCenterAskDao = new HelpCenterAskDao();
+                helperCenterAskDao.findById(req.params.id, (err, resultHelperCenterAsk) => {
                     if (err) {
-                        return resp.status(500).send(JSON.stringify({ error: 'Não é ADMIN' }));
-                    } 
-                    const helperCenterDao = new HelperCenterDao();
-                    helperCenterDao.findById(req.body.id_helpCenter, (errorHelper, resultHelper) => {
-                        if (!resultHelper) {
-                            console.log(errorHelper)
-                            return resp.status(400).send(JSON.stringify({ erro: 'HelpCenter não existente' }));
-                        }
+                        return res.status(400).send(JSON.stringify({ erro: "Houve Algum problema na hora de mostrar os dados da daily favor olhar o log" }));
+                    }
+                        if(resultHelperCenterAsk.id_user == req.body.id_user){
+                            const helperCenterDao = new HelperCenterDao();
+                                helperCenterDao.findById(req.body.id_helpCenter, (errorHelper, resultHelper) => {
+                                    if (!resultHelper) {
+                                        console.log(errorHelper)
+                                        return resp.status(400).send(JSON.stringify({ erro: 'HelpCenter não existente' }));
+                                    }
+                                    helperCenterAskDao.update(req.body, req.params.id, (err, resultUp) => {
+                                        if (err) {
+                                            console.log(err);
+                                            return resp.status(400).send(JSON.stringify({ erro: "Houve Algum problema na hora de atualizar a daily favor olhar o log" }));
+                                        }
+                                        return resp.status(200).send(resultUp);
+                                    });
+                                });
+                        }else{
+                            userDao.checkAdmin(resultByID.email, (err, docs) => {
+                                // console.log(docs.isAdmin);
+                                if (err) {
+                                    return resp.status(500).send(JSON.stringify({ error: 'Não é ADMIN' }));
+                                } 
+                                if (!docs) {
+                                    return resp.status(500).send(JSON.stringify({ error: 'Não é ADMIN' }));
+                                } 
 
-                        const helperCenterAskDao = new HelpCenterAskDao();
-                        helperCenterAskDao.update(req.body, req.params.id, (errorHelperAsk, resultHelperAsk) => {
-                            if (!resultHelperAsk) {
-                                console.log(errorHelperAsk)
-                                return resp.status(400).send(JSON.stringify({ erro: 'Houve Algum problema na hora de cadastrar a daily favor olhar o log' }));
-                            }
-                            return resp.status(201).send(resultHelperAsk);
-                        });
+                                const helperCenterDao = new HelperCenterDao();
+                                helperCenterDao.findById(req.body.id_helpCenter, (errorHelper, resultHelper) => {
+                                    if (!resultHelper) {
+                                        console.log(errorHelper)
+                                        return resp.status(400).send(JSON.stringify({ erro: 'HelpCenter não existente' }));
+                                    }
+
+                                    const helperCenterAskDao = new HelpCenterAskDao();
+                                    helperCenterAskDao.update(req.body, req.params.id, (errorHelperAsk, resultHelperAsk) => {
+                                        if (!resultHelperAsk) {
+                                            console.log(errorHelperAsk)
+                                            return resp.status(400).send(JSON.stringify({ erro: 'Houve Algum problema na hora de cadastrar a daily favor olhar o log' }));
+                                        }
+                                        return resp.status(201).send(resultHelperAsk);
+                                    });
+                                });   
+                            
+                            });
+                        }
+                    
                     });
                 });
-            });
+            }
         }
-    }
 
     list() {
         return (req, resp) => {
@@ -103,7 +132,7 @@ class HelperCenterAskController extends Controller {
                 }
 
                 let response = new Array();
-
+                console.log(result);
                 let docs = result.docs;
                         
                 docs.forEach(doc => {
@@ -141,18 +170,45 @@ class HelperCenterAskController extends Controller {
                 if (result === null) {
                     return resp.status(400).send(JSON.stringify({ erro: 'HelpCenter nao cadastrada' }));
                 }
-                helperCenterAskDao.remove(req.params.id, (error, result) => {
-                    if (error) {
-                        console.log(error);
-                        return resp.status(400).send(JSON.stringify({ erro: 'Houve Algum problema na hora de remover o usuario favor olhar o log' }));
+                const userDao = new UserDao();
+                userDao.findById(req.headers.id_user, (error, resultByID) => {
+                    if (!resultByID) {
+                        console.log(error)
+                        return resp.status(400).send(JSON.stringify({ erro: 'USUARIO não existente' }));
                     }
-                    return resp.status(200).end(JSON.stringify({ msg: 'HelpCenter removido' }));
+                    helperCenterAskDao.findById(req.params.id, (err, resultHelperCenterAsk) => {
+                        if (err) {
+                            return res.status(400).send(JSON.stringify({ erro: "Houve Algum problema na hora de mostrar os dados da daily favor olhar o log" }));
+                        }
+                        if(resultHelperCenterAsk.id_user == req.headers.id_user){
+                            helperCenterAskDao.remove(req.params.id, (error, result) => {
+                                if (error) {
+                                    console.log(error);
+                                    return resp.status(400).send(JSON.stringify({ erro: 'Houve Algum problema na hora de remover o usuario favor olhar o log' }));
+                                }
+                                return resp.status(200).end(JSON.stringify({ msg: 'HelpCenter removido' }));
+                            });
+                        }else{
+                            userDao.checkAdmin(resultByID.email, (err, docs) => {
+                                // console.log(docs.isAdmin);
+                                if (err) {
+                                    return resp.status(500).send(JSON.stringify({ error: 'Não é ADMIN' }));
+                                } 
+                                if (!docs) {
+                                    return resp.status(500).send(JSON.stringify({ error: 'Não é ADMIN' }));
+                                }
+                                helperCenterAskDao.remove(req.params.id, (error, result) => {
+                                    if (error) {
+                                        console.log(error);
+                                        return resp.status(400).send(JSON.stringify({ erro: 'Houve Algum problema na hora de remover o usuario favor olhar o log' }));
+                                    }
+                                    return resp.status(200).end(JSON.stringify({ msg: 'HelpCenter removido' }));
+                                });
+                            });
+                        }
+                    });
                 });
-
-
             });
-
-
         }
     }
